Access Control for Admin Panel

+++ app/Http/Controllers/PostController.php
@@ -9,6 +9,18 @@ use DB;
 class PostController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except'=>['index', 'show']]);
+    }
+
+    /**
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -78,7 +90,12 @@ class PostController extends Controller
     {
          $post = Post::find($id);
 
-         return view('posts.edit')->with('post', $post);
+         // Check for correct user
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized User');
+        }
+        return view('posts.edit')->with('post', $post);
+
     }
 
     /**
@@ -113,6 +130,11 @@ class PostController extends Controller
     public function destroy($id)
     {
        $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized User');
+        }
+
        $post->delete();
 
        return redirect('/posts')->with('success', 'Post Removed');
